requirements.txt
discord
python-dotenv
colorama
art
